Added test for fetching multiple files and data deduplication.
Tests no longer depend on external servers but the connections are instead faked using httpretty.

The multiple fetch test is ugly because it contains a sleep statement instead of faking time change. But we'll fix that some time in the future if the test suit ever becomes large enough that running the tests take a lot of time.